Break out field select into its own component.
